move parse inputs to ui with clearer name, remove traces
